+++ front/app/containers/Admin/pagesAndMenu/containers/ContentBuilder/components/CraftComponents/HomepageBanner/LayoutSettingField/LayoutSettingField.test.tsx
