+++ front/app/containers/Admin/pagesAndMenu/containers/ContentBuilder/components/CraftComponents/Proposals/index.tsx
