+++ front/app/containers/Admin/pagesAndMenu/containers/ContentBuilder/components/CraftComponents/TwoColumn/index.tsx
