+++ front/app/containers/Admin/pagesAndMenu/containers/ContentBuilder/components/CraftComponents/TwoColumn/messages.ts
