+++ front/app/containers/Admin/pagesAndMenu/containers/ContentBuilder/components/HomepageBuilderToolbox/index.tsx
@@ -13,20 +13,20 @@ import SectionTitle from 'components/admin/ContentBuilder/Toolbox/SectionTitle';
 import DraggableElement from 'components/admin/ContentBuilder/Toolbox/DraggableElement';
 
 // widgets
-import TwoColumn from '../CraftSections/TwoColumn';
-import ThreeColumn from '../CraftSections/ThreeColumn';
+import TwoColumn from '../CraftComponents/TwoColumn';
+import ThreeColumn from '../CraftComponents/ThreeColumn';
 import ImageMultiloc from 'components/admin/ContentBuilder/Widgets/ImageMultiloc';
 import IframeMultiloc from 'components/admin/ContentBuilder/Widgets/IframeMultiloc';
 import AccordionMultiloc from 'components/admin/ContentBuilder/Widgets/AccordionMultiloc';
 import WhiteSpace from 'components/admin/ContentBuilder/Widgets/WhiteSpace';
 import ButtonMultiloc from 'components/admin/ContentBuilder/Widgets/ButtonMultiloc';
-import ImageTextCards from '../CraftSections/ImageTextCards';
+import ImageTextCards from '../CraftComponents/ImageTextCards';
 import TextMultiloc from 'components/admin/ContentBuilder/Widgets/TextMultiloc';
-import Proposals from '../CraftSections/Proposals';
+import Proposals from '../CraftComponents/Proposals';
 
 // types
 import { Locale } from 'typings';
-import Events from '../CraftSections/Events';
+import Events from '../CraftComponents/Events';
 
 type HomepageBuilderToolboxProps = {
   selectedLocale: Locale;
